requirements.txt
appdirs==1.4.3
biopython==1.69
Django==1.11
graphviz==0.6
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
Gunicorn
